Restructure asserts
